ponemos guard de rutas y alert de eliminacion accidental

+++ lista_angular/lista/src/app/components/listado-alumnos/listado-alumnos.component.ts
@@ -1,5 +1,9 @@
 import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/firestore';
+import Swal from 'sweetalert2';
+
+
+
 
 @Component({
 selector: 'app-listado-alumnos',
@@ -49,8 +53,19 @@ return false
 }
 
 eliminar(id:string){
-console.log('id', id);
-this.alumnosRef.doc(id).delete()
-}
+  console.log('id', id);
 
+  Swal.fire({
+  title: '¿Estás seguro que deseas eliminarlo?',
+  showCancelButton: true,
+  confirmButtonText: 'Yes, delete it!',
+  cancelButtonText: 'No, keep it'
+  }).then((result) => {
+  if (result.value) {
+  // QUIERO ELIMINAR A ESTE USUARIO;
+  this.alumnosRef.doc(id).delete()
+
+  }
+  })
+}
 }
